Implemented back click on workout_window
Setup the back click on the workout_window. Now it moves back from whatever set you were on and allows you to update that. This makes a more intuitive design in my opinion.

The following still needs to be done:
* Long back click needs to be setup
* All the other windows need to be setup
* More intensive testing should be done as well

Soon we can migrate to a different feature and put that together.