Setup a better model
Added a header file for the model. This will make including it much easier to do. I also redesigned a couple things. We will try to do as little persistence stuff as possible to ensure that we don't run the battery out too much. We will see what we can do.