Added binary decoding function
Added the binary decoding function to get Workout structs from integer values. This needs to be debugged. It looks like I need to pass in array pointers that will store all the reps for each of the different exercises. That will solve the issue, and will allow the user to do everything that he/she needs to.

Maybe I won't go to the gym tomorrow so I can get this working...