Deleting workouts implemented
The model now supports deleting workouts at some index. Deleting automatically updates the rest of the model, so if you delete in the middle of the collection, the rest of the collection will move up to avoid having an empty space.

The model is now almost finished; I think I have just a couple more things to look at and put together and then I'll be done with it.