Tried SimpleMenuLayer for log_menu_window
I tried out using the SimpleMenuLayer for the log_menu_window. It doesn't work. Like at all. We need an array of variable length for it to work, and that just can't be done (simply) on pebble. So we're going to go for something else. Maybe a regular MenuLayer will work. We will try it out.