Added the old_entry_window
Added the old_entry_window. The app currently crashes when you attempt to access it. I'm not entirely sure why. It seems to occur when it tries to get pushed onto the window stack. Maybe the stack is too big? Then again, only three windows... That shouldn't be too bad...

It could be due to the window being null when it's not supposed to, or it being deleted before having an opportunity to get pushed. Further research is required.