Created old_entry_window
This time it's for real though. I've built and rebuilt and rebuilt it again over and over and I finally was able to get it to work. There's still some kinks I need to figure out (like printing out 5 5 5 5 5 for Deadlifts) but other than those things, the old_entry_window is pretty much done. Here's a list though of what I need to work out:

* Deadlifts should only show "5" or maybe "5 x 1" on the old_entry_window
* Holding the select button should show a window asking if you want to delete that workout
* Up and down clicks need to be properly hooked up and tested

This app is coming along really nicely. I'm excited to see it when it's done.