Added deadlift logic for old_entry_window
Added logic for deadlifts to show up properly on the old_entry_window. All we need to do know is the setup_workout_window, nuke the whales window, and all of that jazz. At this rate, we may be done with it for SDK 2.x by the end of the week!