Fixed memory issues
No memory leaks should occur anymore. I changed the way that memory is handled internally, so it should only destroy all of the windows when the app leaves.

This has been a lot of fun, and I'm pumped to release it to the store soon. I hope Mehdi is cool with my app.