Finish refactoring workout_window
Ran into some weird issues with saving the year of a workout. I couldn't remember how it was supposed to be stored, so I tinkered around until I got it to work. I had some weird behaviors with the old_entry_window. I threw in a workaround, but it doesn't seem to match up with the other stuff.

C is a weird language.